+++ app/Models/Post.php
@@ -2,37 +2,10 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
 
-class Post
+class Post extends Model
 {
-    private static $blog_posts = [
-        [
-            'title' => 'sapa',
-            'slug' => 'sapa',
-            'author' => 'budi',
-            'body' => 'hallo apa kabar ?',
-        ],
-        [
-            'title' => 'balas',
-            'slug' => 'balas',
-            'author' => 'dimas',
-            'body' => 'baik, wkwk',
-        ],
-        [
-            'title' => 'konfirmasi',
-            'slug' => 'konfirmasi',
-            'author' => 'firman',
-            'body' => 'jam sekolah mau mulai nih ?',
-        ],
-
-    ];
-
-    public static function all() {
-        return collect(self::$blog_posts);
-    }
-
-    public static function find($slug) {
-        $posts = static::all();
-        return $posts->firstWhere('slug', $slug);
-    }
+    use HasFactory;
 }
